Fix allowed pylon position points from neighbour tiles
Commit 26035c6f3d799a83f22cfdea72529ab8742e87fa changed the way
in which catenary wires were drawn, to omit wires that did not
connect to electrified tracks on a neighbour tile, but introduced
a bug in the handling of allowed and preferred pylon position
points, because it messed up the pylon control point direction to
use when dealing with tracks inciding from a neighbour tile when
there was a track but no wire to be drawn. This was partially fixed
in 215bcb7d2741bf7577ffa9527ebbaa43dc21f90a for the preferred pylon
position points, but not for the allowed pylon position points.
Complete the fix now.