Use TileAreaIterator in CmdConvertRail
Use the new TileAreaIterator implementations in CmdConvertRail,
so that using a rotated area is encapsulated in one struct, and
no orthogonal tile area is used. This fixes a bug that triggered
an assertion because of mismatched costs between test and exec
runs when converting rail over a rotated area that included a
bridge or tunnel head whose counterpart was also in the area
but not in the (non-rotated) rectangle determined by the area
endpoints.