Rework DumpTarget for YAPF debugging output
The way in which YAPF state was output for debugging was just
plain insane. The code would construct one huge string in memory
for each of two pathfinders and them dump them whole to one file
each, without even caring to handle them sequentially. Instead,
we now have the dump code directly write the data to a file as
they are generated.