Add support for deleted functions
Add support for deleted functions through a new DELETED macro,
which expands to a deleted definition for recent versions of GCC
and to an unconditional runtime assertion otherwise. This will
catch attempts to use such functions at compile time at least on
some configurations. The copy constructor and assignment operator
in ThreadMutexLocker are converted to use this new macro.