Introduce a tile set implementation
Introduce a new implementation of a tile set, which keeps track
of a set of objects on the map, each one assigned to a tile, and
allows to test the existence of such an object in the vicinity
of a tile, or to find the object that is closest to a given tile,
without having to check every object in the set. This will be
used to keep sets of towns and industries.