Fix finding of nearest ship depot
When a ship is sent to a depot, it would always try to go to the
depot that is nearest location-wise, which may or may not be the
nearest one path-wise. In fact, if that nearest depot is not
reachable (for instance, if it is landlocked), the ship would be
reported as lost, even if other depots were within reach. This
is now fixed for YAPF, with the addition of a new function to
look for the nearest ship depot that takes pathfinding into
account.

The original pathfinder remains broken, but trying to fix it
is not really worth the while.