Prune the right branch on red signal end-of-line
With yapf.rail_firstred_twoway_eol set to true, when the first
signal found is a red two-way signal, the path leading to it
must be pruned in case the best intermediate node is an earlier
node in the branch but without an intervening junction, because
otherwise heading for that intermediate node would eventually
lead to the red two-way signal anyway. However, the code in
PruneIntermediateNodeBranch was pruning the branch containing
the best intermediate node irrespectively of whether it was an
ancestor of the segment with the red two-way signal or there was
a junction in-between. Fix this by pruning the branch only when
appropriate, and inline PruneIntermediateNodeBranch while at it.