Fix trains getting stuck when another train is close behind
Make FollowReservation return whether there was a reservation beyond
the initial tile, and use this knowledge in FollowTrainReservation
to refrain from looking for a train on the reservation ending
position. Otherwise, if the train has no reservation ahead, such a
search can find another train behind the current one but on the
same position (if they are both in the same tunnel, or in the same
tile with careful planning) and then TryPathReserve will think it
cannot extend the reservation and the train will be stuck forever.