Move CYapfRailBaseT::AddTargetCost to CYapfRailOrderT
Add a new function CYapfRailOrderT::AddTargetCost with the contents
of CYapfRailBaseT::AddTargetCost, and leave the latter empty, so
that only trains heading for a station (using CYapfRailOrderT) will
have their target node cost adjusted. This means that trains merely
looking for a safe tile (through CYapfAnySafeTileRailT) will not get
any special penalty or bonus if the tile found belongs to a station.