Treat complex waypoint penalty as a target penalty
Target complex waypoint tiles may be given an extra penalty
depending on whether they are reserved or not, to make unreserved
waypoint tiles preferred as target nodes. Add this penalty as a
target penalty, not as a segment penalty, which is the right thing
to do. As a consequence, we no longer need to disable the cache
when the destination is such a waypoint.