Disable the cache when masking reserved tracks in YAPF
As it is, the rail pathfinder only ever marks a segment as ending
in a safe tile if it has been constructed to mask reserved tracks.
This means that masking of reserved tracks drastically changes the
definition of a segment: while normally a segment is a stretch of
track that can contain multiple signals, enabling masking makes
segments end at the first signal. Thus, segments cannot be used
interchangeably with different values of mask_reserved_tracks.
To make up for this, CYapfRailBaseT::TryReservePath invalidated the
cache when a reservation ended in a cached node (note that this is
very, very unlikely to happen, if at all, because there is a safe
tile at or before any signal, so a reserved path could only ever
end in a cached node if the value for rail_look_ahead_max_signals
is reduced from its default 10 to its minimum 1). But this is too
broad a hammer, even if unlikely; instead, disable the cache when
masking reserved tracks, so pathfinding in that case cannot thrash
the cache.