Improve sanity checking for YAPF cache
Add a new sanity check for YAPF cached rail segments when debugging.
Instead of running the pathfinder twice and only comparing the final
results, check at every node that the cached data match a fresh
computation. This will detect all cache incoherencies, and not just
those that affect the final outcome, and will do so at the earliest
possible opportunity.