Reintroduce a maximum cost flag in enum EndSegmentReason
Reintroduce a maximum cost flag in enum EndSegmentReason, named
ESR_MAX_COST, to be used when a node exceeds the maximum cost set
for a path. This was removed (under the name ESR_PATH_TOO_LONG)
because it was the only flag that had to be masked away when
storing a cached segment, but now there are provisions in the
code to ensure that ESR_MAX_COST is never added to a cached node.