Allow rail YAPF to use the cache when maximum cost is set
Allow rail YAPF read-only access to the cache when a maximum cost
is set. This means that CYapfRailBaseT::CalcNode will look for the
segment in the cache and use it if present and below the maximum
cost; otherwise, it will create a fresh, non-cached segment, so
that an interrupted segment cost computation will not get cached.