Simplify StringIterator definition with respect to ICU
Instead of having a base class StringIterator full of virtual
methods that had exactly one derived class on any given compilation
(IcuStringIterator or DefaultStringIterator, depending on the value
of WITH_ICU), make it a typedef for the class that is actually
defined. This avoids going through virtual calls for something
that, in practice, is never virtual.