Add a new macro cpp_pointer to get a pointer to a field
Add a new macro cpp_pointer to get a pointer to a given field
in a non-existent object of a given type. Even though the object
does not exist, this is useful when only the type of the pointer
matters or when trying to access a static member of a field that
is itself a class.