Make the custom currency separator a string setting
The only difference between a char setting (SDT_CHR) and a string
setting (SDT_STR) is how they are handled in savegames, but the
custom currency separator is never stored in a savegame. This
removes the only occurrence of SDT_CHR in the code and will also
allow non-ASCII characters as separators.