Unify implementation of TileBelongsToRailStation
Stations and waypoints share the same pool, so checking the station
index of a tile is enough to tell if it belongs to a given station
or waypoint. This commit gives BaseStation an implementation of
TileBelongsToRailStation, and makes it non-virtual, which avoids
indirect function calls. We keep the implementations in derived
classes Station and Waypoint, since they agree with that of the
base class and, as long as rail stations and waypoints have
different station type ids, checking for a particular one should
be faster than checking for either of them.