Rework handling of station sign updates
Make UpdateStationSignCoord always call UpdateVirtCoord, rename
it to UpdateStationSign to better reflect what it does and always
invoke it directly from the functions that remove station parts,
before RecomputeIndustriesNear, instead of hooking into it from
DeleteStationIfEmpty, so that RecomputeIndustriesNear already has
an up-to-date station sign. This fixes a crash when removing a
station part that leaves the station sign momentarily outside of
the station area.