Use awk instead of cpp to generate the extra grf nfo file
(openttd r26708) -Codechange: replace C preprocessor with all kinds
of options to prevent adding certain things to the result and the
sed script to remove everything that shouldn't be in the resulting
nfo file but came out of the C preprocessor by a small-ish awk
program. This means no more breaking if a C preprocessor decides
to add more data to the processed file (like GCC's inclusion of
stdc-predef.h when not passing -nostdinc). It furthermore more
than halfs the time needed for the assemble step of the extra GRF
compilation