Overhaul implementation of SaveLoad construction macros
Completely overhaul the implementation of the SaveLoad construction
macros (SLE_*): They no longer use SLE_GENERAL or SLEG_GENERAL
helpers (but use other helpers instead), and handling of global
objects is closer to that of struct fields. This change affects
the settings construction macros; SDT_GENERAL and SDTG_GENERAL
are removed in the process.