Use an orthogonal iterator in TryBuildTransmitter
A simple orthogonal tile iterator is cheaper to use than a circular
search when the area iterated over is a square, and the result is
the same if we are only interested in whether a tile satisfying
certain conditions exists, and not in the first one that does.

While at it, inline HasTransmitter, which was only used there.