Have RiverModifyDesertZone take care of the conversion area
Have RiverModifyDesertZone iterate around a given tile, removing
desert status from the area iterated over, instead of being used
as a callback to CircularTileSearch. Also use a simple orthogonal
tile iterator, since we are iterating over all the tiles in the
square anyway and the order in which we do so does not matter.