Introduce CommandSource for the source of a command
Introduce a new enum CommandSource to store the source of a command.
This allows us to track the origin of a command and whether it has
already been sent and received over the network, if applicable.
This will be useful in removing the callbacks from the command
structure. Note that some of the values are not used yet.