Remove callback argument from DoCommandP
With the recent changes to callbacks and the addition of a command
source argument to DoCommandP, it is no longer necessary to pass it
a callback argument, since DoCommandP can determine what callback
to call on its own. This also means that, as a bonus, the callback
does not need to be sent across the network and the callback list
in the network command code can be removed.