New order clone flag CO_UNSHARE_COPY
Add a new order cloning flag CO_UNSHARE_COPY to be used with
CMD_CLONE_ORDER when taking a vehicle out of a shared order group.
It is semantically equivalent to CO_COPY, but it helps understand
what is meant and will make it easier to determine the message to
show on error.