Remove wrong assertion when reversing in the track follower
When the track follower considered reversing a non-tram road vehicle
at end of line, it asserted that the last (old) tile always had some
valid trackdirs after reversing, but this may not be the case, if
the last tile is a one-way road. Remove the assertion and handle
this case.