Move most of TextfileWindow::LoadTextfile into its constructor
Move most of TextfileWindow::LoadTextfile into its constructor;
the constructor now takes some extra arguments and is in charge
of loading the textfile, sanitising its contents and arranging
it into lines. The only part that is not moved is the check for
missing glyphs in the current font, which is now in a new method
TextfileWindow::CheckForMissingGlyphs. This is left as a separate
method because it triggers a storm of updates in all registered
windows, including the TextfileWindow under construction, so it
cannot be called until the deriving superclass of TextfileWindow
has finished initialising itself.