Properly set the timetabled flag on old current orders
Beginning with savegame version 19, the timetabled status for
travel/wait time in an order is kept as a separate bit. There is
saveload code to set these bits when loading an older savegame,
but this code only ran on orders in a list, and not on the copy
that each vehicle maintains as its current order, which was
causing certain orders to be untimetabled on load. Fix this by
applying the order status adjustment also to current orders.