Allow to forbid short turns in controller_front_new_tile
Add a parameter to controller_front_new_tile to tell it not
to consider doing a short turn at the tile edge but to avoid
reversing if possible or to do a long turn otherwise instead.
This helps when exiting a standard road stop and fixes a bug
when a road vehicle tries to exit a wormhole but the wormhole
head only has the road piece that heads back into the wormhole.