Assert that articulated vehicles do not enter standard stops
RoadStop::EnterStandard was checking if the vehicle trying to enter
the stop was an articulated vehicle, and returning false in that
case. However, if that code path had been ever reached, the vehicle
would have become stuck forever in controller_standard_stop,
additionally blocking the road stop entrance. In fact, articulated
road vehicles never even attempt to enter a standard road stop
because of provisions in RoadChoosePath to this effect, so this
dead check has been removed.