Do recompute station areas when loading old savegames
AfterLoadStations was in charge of recomputing the station area of
the bus stop, truck stop and dock parts of a station, but was only
called for savegame versions earlier than 27. Fix this by taking
out the update of station speclists into its own function and
merging the rest with AfterLoadRoadStops.