Fix removal of RVS_IS_STOPPING flag from TTO/TTD road vehicles
Commit 533ddeb9e6352774b441e4afcc3f60125bba74db tried to allow
loading of some TTDP savegames that used the RVS_IS_STOPPING
bit for road vehicles, clearing that bit on load, but did so
unconditionally, which is wrong. This was partially fixed in
commit a84e6b015d2715daa872f11f1f3893ce06e9f458 for wormholes
and depots, but it was still cleared for other vehicles not in
a station, which is wrong, too. This bug would manifest itself
as a vehicle that jumps once or twice in its tile and then
recovers itself, if it happened to be turning left or right
at the moment of the save. However, after the road vehicle
controller rewrite, which has stricter checks, this loading
bug made an assertion fail.