Compare speeds using the same units in the road pathfinder
The road pathfinder was comparing the maximum speed of a bridge
in internal road vehicle units with the maximum speed of a vehicle
in display units; as a consequence, bridges with a maximum speed
greater than half the maximum speed of the pathfinding vehicle
would not impose a speed penalty. Fix this now.