Add another function to free the reservation of a train
Make the exported variant of FreeTrainTrackReservation return
whether the given train had any reservation at all (not heading
into a depot) and use this knowledge in a new function
FreeTrainReservation that finds the train that reserved a certain
track and returns that train only if it did had a reservation
other than the track underneath, in which case it also frees
that reservation. This fixes a crash when modifying signals on a
track on which there is a train that is partially in a depot.