Use a single byte to send the command id over the wire
Now that the command field of a Command stores only a command id
(and not any flags or an error string), a single byte is enough
to send it over the wire. While at it, reduce the width of the
command field in desync debug output (2 hex chars instead of 8).