Check town rating earlier when building a dock
Check if the local authority allows to build a dock before trying
to clear the landscape. The order matters here: if we try to clear
the landscape first, the command may succeed during the test run
(because no changes are made to the landscape) but later fail during
the exec run (because clearing a tile removed some trees and that
made the town rating drop below the point at which building new
stations is allowed).