Fix ttd_unique_free_ptr emulation through ttd_shared_ptr
The constructor of ttd_unique_free_ptr, when using ttd_shared_ptr,
lacked the right deleter in the constructor. Also, we add a proxy
function for free to be used as deleter, since using free directly
seems to confuse MSVC because the cast required for overload
resolution apparently messes up the calling convention.