Rework news creation system
Rework the news type system into a hierarchical tree of classes,
each with its custom constructors. This avoids the need to use
the global string parameter stack to pass the parameters to news
items on construction, and also fixes memory leaks when a news
item would be generated but is suppressed while in the main menu.