Fix a crash when removing a path signal under a train
Removing a path signal under a train that was partially into a depot
would cause a crash. This is now fixed by partially reverting commit
a73366305fda6414907c9ad76168c4b88a06a183 and delegating the task of
freeing the reservation to GetTrainForReservation, through a new
parameter. Also, GetTrainForReservation will now return NULL if the
train that holds the reservation is partially into a depot, since
then the only actual reservation is the stretch of track under the
train itself, which can be neither freed nor extended, and no caller
of GetTrainForReservation cares about that case.