Replace AIInfo::use_as_random with AIInfo::use
Replace boolean field AIInfo::use_as_random with a new enum field
AIInfo::use that can take three different values, according to
whether the AIInfo instance represents a script that can be used
as random script, cannot be used as random script, or is the dummy
script. This will be useful in next commit.