Rework the squirrel printing system
Rework the squirrel printing system so that the print function used
by class Squirrel takes a squirrel virtual machine as an additional
first argument, and a format string and a va_list argument instead
of an already formatted message. This is required to remove class
ScriptController in a later commit.