Move script register functions to core script files
Move the function that registers the AI scanner API to a Squirrel
engine to ai_core.cpp, along with the function that eventually
registers the script to the script list, and similarly for game
scripts. This allows them to be ordinary static functions instead
of methods in the AI/GS Info/Library classes.