Remove both ends of a rail bridge before updating signals
Fully remove both bridge heads before adding anything to the signal
buffer in RemoveRailBridge. Otherwise, if you were really unlucky
and an automatic update of signals triggered in between, while one
of the bridge heads still existed on the map, the signal buffer code
would crash when looking for the other one.