Add a new struct FontMetrics to cache some font metrics
Add a new struct FontMetrics to hold a cache of some metrics of
a font, to be used for _stringwidth_table, which is renamed to
font_metrics_cache in the process. While this is only used for the
widths of ASCII characters for now, it will grow extra fields in
next commit.