Fix advertisement period in dedicated servers
(openttd r27400) -Fix [FS#6368] (r26449): when a dedicated
server was paused with no clients the tick length was increased
significantly, making any assumptions about the tick length
used further down in the code are not true anymore. One of such
assumptions was that one should readvertise every 15 minutes worth
of original ticks, but due to the lengthening this timeframe would
be more like 45-60 minutes. Now we'll take the operating system's
millisecond counter instead