Rearrange house sprite data array
Rearrange _town_draw_tile_data to make it a 3-dimensional array
of pointers to DrawBuildingsTileStruct, and rename it out of the
reserved namespace. This makes data access saner and reduces both
the number of lines of code and the size of the resulting binary,
as it takes advantage of data duplicity.

   text    data     bss       dec     hex filename
6496927   70880 4051440  10619247  a2096f bin/openttd (before)
6475135   70880 4051440  10597455  a1b44f bin/openttd (after)