Rearrange industry sprite data array
Rearrange _industry_draw_tile_data to make it a 2-dimensional array
of pointers to DrawBuildingsTileStruct, and rename it out of the
reserved namespace. This makes data access saner and reduces both
the number of lines of code and the size of the resulting binary,
as it takes advantage of data duplicity.