Add a new struct FakeHouseResolverObject
Add a new struct FakeHouseResolverObject to provide a resolver
object for houses that are not constructed yet (for instance,
to draw them in the GUI), instead of HouseResolverObject. Also
add a new overloaded variant of GetHouseCallback that uses this
new object and does not require a tile or a town.