Always try to honour the autorefit cargo set
If a vehicle has an autorefit order but none of the chosen cargoes
are available to load at the station, the vehicle is not refitted
and keeps its previous cargo type. However, this happened even if
the previous cargo type was not one of the allowed refits, which
means that the vehicle might end up loading some unintended cargo.
Fix this by trying to refit the vehicle to one of the cargoes in
the refit set in this case.