Add a helper struct GenericScopeResolverData
Add a new struct GenericScopeResolverData to hold the data required
by callback CBID_GENERIC_AI_PURCHASE_SELECTION, and keep a pointer
to it in GenericScopeResolver. This allows to remove the ai_callback
field, which now is equivalent to doing a NULL check on data, and
will allow for further changes.