Fix FindNearestDepotOrigin for reversing vehicles
When FindNearestDepotOrigin was called for a reversing vehicle,
it would blindly assume that the vehicle would drive on a straight
trackdir after reversing, which was wrong for two reasons: Firstly,
because the straight trackdir may not be available on the tile (for
instance, after a short turn on a tile missing the opposing road
piece) and, secondly, because reversing vehicles get to call
the pathfinder after the turn, so they are not forced into any
particular trackdir. Fix this by using all available trackdirs
on the tile for the initial node.