Call Astar::InsertOpenNode only when adding new nodes
Astar::InsertOpenNode was called from ReplaceNode, InsertInitialNode
and InsertNode, but the call from ReplaceNode was different in that
the node with which it was called was not a newly created one.
Inline the call to InsertOpenNode from ReplaceNode and leave
InsertOpenNode proper for calls from InsertInitialNode and
InsertNode, and add some assertions along the way.