Change the signature of Window::OnPlacePresize
Change the signature of Window::OnPlacePresize so that it receives
a couple of tile pointers, both pointing to the tile where the mouse
is initially, and the function is expected to set those tiles to the
starting and ending tiles of the tile area to select. Calling
VpSetPresizeRange is then done in HandlePlacePresize, which is the
only user of Window::OnPlacePresize.