Remove HT_DIAGONAL from enum HighLightStyle
Whether to allow rotated selections or not is a property of the drag
mode, and not of the cursor mode. This commit removes HT_DIAGONAL
from enum HighLightStyle and adds a new value VPM_X_AND_Y_ROTATED
to enum ViewportPlaceMethod to be passed to VpStartPlaceSizing when
rotated selections are allowed.