Remove the last parameter of ReadSprite
GetRawSprite is only called from GetSprite, GetNonSprite
and GetMapGenSprite in spritecache.h; and ReadSprite and
HandleInvalidSpriteRequest in spritecache.cpp. Calls to GetRawSprite
from GetSprite, GetNonSprite and GetMapGenSprite always use NULL
for the allocator parameter, while calls from ReadSprite and
HandleInvalidSpriteRequest use the allocator argument received
from GetRawSprite itself, which is their only caller, and this
argument is set to either GetRawSprite's own allocator argument
or to AllocSprite. This means that GetRawSprite is only ever invoked
with either AllocSprite or NULL for its allocator argument, and then
ReadSprite is always called with AllocSprite as allocator.