Fix compilation without network support
Including network_content_gui.h from strings.cpp to get the
declaration of BuildContentTypeStringList causes compilation to fail
when networking is disabled because ContentInfo is not defined. Fix
this by wrapping the contents of network_content_gui.h in an #ifdef
block so that only a dummy version of BuildContentTypeStringList
is defined when networking is disabled.