Add a new file with a flexible array helper class
Add a new file core/flexarray.h with the definition of a helper
class template to be used by classes that have flexible array
members, to delete the copy constructor and assignment operator
and to implement an operator new with a different signature.
Also use this new base class in all three clases that already
have flexible array members.