Remove Textbuf::MovePos
Remove Textbuf::MovePos and replace it with Textbuf::MoveLeft,
Textbuf::MoveRight and Textbuf::MoveEnd, which handle moving
left, right and to the end of the text; handling of the home
key is inlined into Textbuf::HandleKeyPress. This is simpler
for the on-screen code, which was the only external user of
Textbuf::MovePos.