Tweak autorenew behaviour with breakdowns disabled
This complements commit ae3640378413cfbbbb6cbe2d87f68f97d570856f in
that it allows the user to decide whether vehicles should look for
a depot on their own for autorenew when breakdowns are disabled.
Additionally to the old possibilities of enabling or disabling
servicing when playing with no breakdowns, there is now a third
option to allow automatic servicing just for autorenew. Also, the
setting is now per-company.