Make DoZoomInOutViewport work on a viewport
Change the signature of DoZoomInOutViewport to work on a viewport
and not on a window. It also now assumes that the new zoom level
is within the allowable bounds, so it no longer returns a boolean
success status, and does not do any window invalidation either,
which is left to its callers.