Do not call InvalidateData in ClampViewportZoom
Do not call InvalidateData at the end of ClampViewportZoom, and
let its callers do it as required, as one of them is doing it
unconditionally anyway so only the other one needs to do it.
As a consequence, ClampViewportZoom no longer needs a pointer
to the window and passing a pointer to the viewport is enough.