Change signature of operator new in FlexArray
Struct FlexArray has a helper placement operator new to make
allocations with extra storage for the flexible array member,
but its signature makes it match a usual (non-placement) operator
delete under C++14, which is not allowed. Change its signature
to have an additional parameter with a default argument to work
around this limitation.