Add overloaded functions GetWireBase and GetPylonBase
Add overloaded functions GetWireBase and GetPylonBase that work
on a RailtypeInfo instead of on a tile and track, and use them in
DrawCatenaryOnBridge. This fixes a bug that could cause custom
pylon/wire sprites for a railtype to be ignored on a bridge if the
bridge head had two different railtypes.