Fix crash on custom rail bridge head track removal
AddCrossingToSignalBuffer did not check if the tile passed in was
a bridge head, and blindly added all four sides of the tile to the
set of tile sides to check, but this triggered an assertion in
UpdateSignalsInBuffer as the bridge side of a bridge head is not
supposed to be in that set. As a consequence, any attempt to remove
the track that is perpendicular to the bridge from a bridge head
containing just both orthogonal tracks would crash the game.