Fix non-stop behaviour for nearest depot orders
An order to visit the nearest depot is converted to an order
to visit the particular depot that is actually closest when it
becomes the current order for a vehicle, but the code would
clobber the non-stop flags for the order when doing so, because
BaseOrder::MakeGoToDepot would be called with the very same order
that it has to change as reference parameter. Fix this by replacing
that BaseOrder::MakeGoToDepot variant with a new method to set the
depot destination for an existing depot order.