Parallelise loading even with autorefit or full load
When a train is waiting for a full load at a station, incoming cargo
was being loaded into the wagons sequentially, i.e., cargo would
be loaded into the first wagon that could take it and only start
filling the second one when the first one was full (or had reserved
enough cargo for a full load), and so on. This meant that the batch
of cargo that finally fills the train would eventually be loaded
into the (mostly empty) last wagon, possibly taking several loading
steps, while all the other wagons were already full. We change this
behaviour so that incoming cargo is now distributed among all wagons
that can take it, balancing their loads, so that the final batch of
cargo can be spread over the remaining capacity of several wagons in
one loading step and the train can leave sooner.