Rework handling of missing glyphs in FontCache
Add a field missing_sprite in FontCache to hold a pointer to the
sprite to be used for missing glyphs (which is either the question
mark of the font or the builtin question mark), and use position 0
in the sprite map only for the builtin question mark sprite if it
is required. This allows to remove the duplicate field from
GlyphEntry.