Allocate 32bpp animation buffer based on surface pitch
Access to a pixel in the 32bpp blitters animation buffer is done
relative to its position in the blitting surface buffer, but the
animation buffer size was computed based on surface width, not
pitch, so out-of-bound accesses were possible if the surface pitch
was strictly greater than its width. Fix this by allocating an
animation buffer of the same size as the surface buffer.