Combine calls to MarkDirty in ScrollScreenRect
Inline the first part of GfxScroll into ScrollScreenRect, which
is its only caller, and add a call to VideoDriver::MarkDirty for
the whole region at the end of ScrollScreenRect. With this, all
calls to RedrawScreenRect in ScrollScreenRect can be simplified to
DrawOverlappedWindowForAll, and the call to VideoDriver::MarkDirty
in GfxScroll can be removed. This means fewer calls to
VideoDriver::MarkDirty (one, instead of up to three) and will
allow for further cleanup