Use YAPF rail segment cache when cost-limited
The comment at the top of CYapfRailBase::CalcNode stated that the
intent was to use of the segment cache if a maximum cost was set
for those segments that were already cached and below that cost,
but the code unconditionally disabled the cache whenever there was
a maximum cost. The code has been changed to match the intent.