Remove ground palettes from house drawing data
Only a few house views had a palette other than PAL_NONE for their
ground sprite, and in all but one cases it had no effect, as the
ground sprite did not have any pixels affected by the recolouring.
This does change a few pixels in the garden part of ground sprite
0x061e, but the original behaviour can be construed as an unintended
side effect, as it did not happen with sprite 0x0620, which is the
alternate view of the same house id. This commit is separate from
the next one to isolate this change.