Add a simpler SplitGroundSpriteForOverlay variant
SplitGroundSpriteForOverlay is only called from two places,
in DrawStationTile and DrawTile_RailStation, but the call
from DrawStationTile passes an explicit null pointer for
the TileInfo parameter, which means that the last part of
SplitGroundSpriteForOverlay is skipped. Extract the part of
SplitGroundSpriteForOverlay that is used from both places into
an overloaded variant and export it instead of the old one.