Rework ReadSpriteLayoutSprite not to work on a ByteReader
Move reading of the raw sprite and its flags, if any, out of
ReadSpriteLayoutSprite and into its callers, rearrange its
parameters as some of them are no longer necessary and rename
the resulting function to AdjustSpriteLayoutSprite. In particular,
the new function no longer operates on a ByteReader buffer, which
will be useful shortly.