Rework NewGRFSpriteLayout not to own its data arrays
Rework NewGRFSpriteLayout not to own the sprite sequence and
register array it points to; instead, its caller/user is responsible
for managing these arrays. Now struct TileLayoutSpriteGroup uses
the equivalent of a flexible array to keep its sprite and register
arrays and no longer derives from ZeroedMemoryAllocator, as a
consequence, and there is a new struct StationTileSpriteLayout
with a similar behaviour for station tile sprite layouts.