Improve drag and drop interface for custom stations
When laying out a station that only allows a fixed number of
tracks or platform length with drag and drop, the particular
dimension that is fixed is now preset in the highlight area
and dragging will only be possible in the other direction.
If both the number of tracks and platform length are fixed,
then the highlight area is set to the only possible size and
the station is built on click without any dragging at all.