Rework airport layout rotation storage
Store the rotation of each airport layout as an extra field in the
first tile of the layout, instead of in an independent array. This
does not increase the size of AirportTileTable, improves airport
tile table data locality, reduces memory allocations when a new
layout is defined and in fact fixes a memory leak, as the rotation
array of an airport spec was not being freed in ResetCustomAirports
if it had been allocated in AirportChangeInfo.