Change GetTrainStopLocation to GetDistanceToStopLocation
Replace GetTrainStopLocation, which was used to compute the
stopping location for a train at a station, with a new function
GetDistanceToStopLocation that computes the remaining distance to
that location instead. The new function takes the current position
of the train within its tile into account, and so provides tile
subcoord accuracy. Also make the function static while at it.