Fix catenary drawing in depots
The wrong sprite was being drawn for the catenary in electrified
depots, which was noticeable if the transparency options were
set to draw the wire but not the depot building. To fix this,
DrawRailDepotCatenary has to invert the direction of the depot
when passing it to DrawRailTunnelDepotCatenary, as the direction
of a tunnel is that into the tunnel but the direction of a depot
is that out of the depot.